--- a/pkg/cacheutil/memcached_server_selector.go
+++ b/pkg/cacheutil/memcached_server_selector.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	"github.com/bradfitz/gomemcache/memcache"
-	"github.com/cespare/xxhash"
+	"github.com/cespare/xxhash/v2"
 	"github.com/facette/natsort"
 )