--- a/pkg/query/querier.go
+++ b/pkg/query/querier.go
@@ -327,7 +327,7 @@ func sortDedupLabels(set []storepb.Series, replicaLabels map[string]struct{}) {
 	// With the re-ordered label sets, re-sorting all series aligns the same series
 	// from different replicas sequentially.
 	sort.Slice(set, func(i, j int) bool {
-		return labels.Compare(labelpb.LabelsToPromLabels(set[i].Labels), labelpb.LabelsToPromLabels(set[j].Labels)) < 0
+		return labels.Compare(labelpb.ZLabelsToPromLabels(set[i].Labels), labelpb.ZLabelsToPromLabels(set[j].Labels)) < 0
 	})
 }