--- a/pkg/receive/handler.go
+++ b/pkg/receive/handler.go
@@ -4,10 +4,11 @@
 package receive
 
 import (
+	"bytes"
 	"context"
 	"crypto/tls"
 	"fmt"
-	"io/ioutil"
+	"io"
 	stdlog "log"
 	"net"
 	"net/http"
@@ -279,20 +280,21 @@ func (h *Handler) handleRequest(ctx context.Context, rep uint64, tenant string, 
 	return nil
 }
 
-var tmp int
-
 func (h *Handler) receiveHTTP(w http.ResponseWriter, r *http.Request) {
 	span, ctx := tracing.StartSpan(r.Context(), "receive_http")
 	defer span.Finish()
 
-	// TODO(bwplotka): ReadAll grow slice a lot, we may want to get content size from request header.
-	compressed, err := ioutil.ReadAll(r.Body)
+	compressed := &bytes.Buffer{}
+	if r.ContentLength >= 0 {
+		compressed.Grow(int(r.ContentLength))
+	}
+	_, err := io.Copy(compressed, r.Body)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	reqBuf, err := snappy.Decode(nil, compressed)
+	reqBuf, err := snappy.Decode(nil, compressed.Bytes())
 	if err != nil {
 		level.Error(h.logger).Log("msg", "snappy decode error", "err", err)
 		http.Error(w, err.Error(), http.StatusBadRequest)