--- a/pkg/rules/manager.go
+++ b/pkg/rules/manager.go
@@ -63,8 +63,8 @@ func (g Group) toProto() *rulespb.RuleGroup {
 					Name:                      rule.Name(),
 					Query:                     rule.Query().String(),
 					DurationSeconds:           rule.HoldDuration().Seconds(),
-					Labels:                    storepb.LabelSet{Labels: labelpb.LabelsFromPromLabels(rule.Labels())},
-					Annotations:               storepb.LabelSet{Labels: labelpb.LabelsFromPromLabels(rule.Annotations())},
+					Labels:                    storepb.LabelSet{Labels: labelpb.ZLabelsFromPromLabels(rule.Labels())},
+					Annotations:               storepb.LabelSet{Labels: labelpb.ZLabelsFromPromLabels(rule.Annotations())},
 					Alerts:                    ActiveAlertsToProto(g.PartialResponseStrategy, rule),
 					Health:                    string(rule.Health()),
 					LastError:                 lastError,
@@ -77,7 +77,7 @@ func (g Group) toProto() *rulespb.RuleGroup {
 				Result: &rulespb.Rule_Recording{Recording: &rulespb.RecordingRule{
 					Name:                      rule.Name(),
 					Query:                     rule.Query().String(),
-					Labels:                    storepb.LabelSet{Labels: labelpb.LabelsFromPromLabels(rule.Labels())},
+					Labels:                    storepb.LabelSet{Labels: labelpb.ZLabelsFromPromLabels(rule.Labels())},
 					Health:                    string(rule.Health()),
 					LastError:                 lastError,
 					EvaluationDurationSeconds: rule.GetEvaluationDuration().Seconds(),
@@ -98,8 +98,8 @@ func ActiveAlertsToProto(s storepb.PartialResponseStrategy, a *rules.AlertingRul
 	for i, ruleAlert := range active {
 		ret[i] = &rulespb.AlertInstance{
 			PartialResponseStrategy: s,
-			Labels:                  storepb.LabelSet{Labels: labelpb.LabelsFromPromLabels(ruleAlert.Labels)},
-			Annotations:             storepb.LabelSet{Labels: labelpb.LabelsFromPromLabels(ruleAlert.Annotations)},
+			Labels:                  storepb.LabelSet{Labels: labelpb.ZLabelsFromPromLabels(ruleAlert.Labels)},
+			Annotations:             storepb.LabelSet{Labels: labelpb.ZLabelsFromPromLabels(ruleAlert.Annotations)},
 			State:                   rulespb.AlertState(ruleAlert.State),
 			ActiveAt:                &ruleAlert.ActiveAt, //nolint:exportloopref
 			Value:                   strconv.FormatFloat(ruleAlert.Value, 'e', -1, 64),