… CHANGELOG.md
… VERSION
… go.sum
… pkg/api/query/v1_test.go
… pkg/query/internal/test-storeset-pre-v0.8.0/storeset_test.go
… pkg/query/querier_test.go
… pkg/query/storeset_test.go
… pkg/receive/handler_test.go
… pkg/receive/hashring_test.go
… pkg/receive/multitsdb_test.go
… pkg/rules/rules_test.go
… pkg/rules/rulespb/custom_test.go
… pkg/rules/rulespb/rpc.pb.go
… pkg/store/bucket.go
… pkg/store/bucket_e2e_test.go
… pkg/store/bucket_test.go
… pkg/store/labelpb/label_test.go
… pkg/store/labelpb/types.pb.go
… pkg/store/labelpb/types.proto
… pkg/store/local.go
… pkg/store/multitsdb.go
… pkg/store/multitsdb_test.go
… pkg/store/prometheus.go
… pkg/store/prometheus_test.go
… pkg/store/proxy.go
… pkg/store/proxy_test.go
… pkg/store/storepb/custom.go
… pkg/store/storepb/custom_test.go
… pkg/store/storepb/prompb/types.pb.go
… pkg/store/storepb/prompb/types.proto