@bwplotka 2020-10-15
Ready for review. I decided to change naming which increased size of this PR a bit...