@bwplotka 2020-10-15
The "leak" removal can be seen on remote write benchmark profiles of inuse space in correct moment ;p 

Before:
![image](https://user-images.githubusercontent.com/6950331/96196206-b7295a00-0f46-11eb-8272-db71a98e9880.png)

After (note, no snappy):
![Uploading image.png…]()