@bwplotka 2020-10-15
Since it's a big change with extra optimizations, let's fix this against master.

Will fix release with quicker fix disabling all yolo strings for now.